add type to While function for testing

+++ while.jl
@@ -1,5 +1,5 @@
 # While loops loop while a condition is true
-function while_func()
+function while_func()::Int
     x::Int = 0
     while x < 4
         println(x)
